# app/Http/Resources/AssociateResource.php
<?php

namespace App\Http\Resources;

use Illuminate\Http\Request;
use Illuminate\Http\Resources\Json\JsonResource;
use OpenApi\Attributes as OA;


class AssociateResource extends JsonResource
{
    public function toArray(Request $request): array
    {
        return [
            'id' => $this->id,
            'name' => $this->name,
            'cpf' => $this->cpf,
            'email' => $this->email,
            'telephone' => $this->telephone,
            'city' => $this->city,
            'state' => $this->state,
            'created_at' => $this->created_at,
        ];
    }
}